Correções para funcionar o cadastro de Convenio (ch, filme)

+++ src/br/bcn/admclin/dao/CONVENIOCH.java
@@ -94,7 +94,7 @@ public class CONVENIOCH {
             stmtInsertValor.setInt(1, convenioCHMODEL.getUsuarioId());
             stmtInsertValor.setDate(2, convenioCHMODEL.getDat());
             stmtInsertValor.setInt(3, convenioCHMODEL.getHandle_convenio());
-            stmtInsertValor.setInt(4, convenioCHMODEL.getValor());
+            stmtInsertValor.setDouble(4, convenioCHMODEL.getValor());
             stmtInsertValor.setDate(5, convenioCHMODEL.getDataAValer());
             stmtInsertValor.executeUpdate();
             stmtInsertValor.close();
